--- gatsby-ssr.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { renderToString } from 'react-dom/server';
-
-import { ContextProviderWithReducer } from './src/context/Overlay';
-
-export const replaceRenderer = ({ bodyComponent, replaceBodyHTMLString }) => {
-  const App = () => (
-    <ContextProviderWithReducer>{bodyComponent}</ContextProviderWithReducer>
-  );
-
-  replaceBodyHTMLString(renderToString(<App />));
-};
