+++ src/components/LanguageList/LanguageListItem.jsx
@@ -3,22 +3,15 @@ import PropTypes from 'prop-types';
 
 import Link from '../Link/LinkGatsby';
 
-import Context, {
-  SET_LANGUAGE_OVERLAY_VISIBILITY,
-} from '../../context/Overlay';
+import OverlayContext from '../../context/Overlay';
 
 // As close as possible to https://github.com/ec-europa/europa-component-library/blob/v2-dev/src/systems/ec/implementations/react/page-structure/language-list/src/LanguageListItem.jsx
 const LanguageListItem = ({ label, lang, href, isActive, isOverlay }) => {
-  const { dispatch } = useContext(Context);
+  const { setOverlayIsHidden } = useContext(OverlayContext);
 
   return (
     <Link
-      onClick={() =>
-        dispatch({
-          type: SET_LANGUAGE_OVERLAY_VISIBILITY,
-          hideOverlay: true,
-        })
-      }
+      onClick={() => setOverlayIsHidden(true)}
       lang={lang}
       hrefLang={lang}
       href={href}
